RadarFII/Docker_RadarFlls/requirements.txt
requests
pandas
matplotlib
python-dotenv
sendgrid
google-generativeai
